backend/requirements.txt
annotated-types==0.7.0
anyio==4.12.0
asgiref==3.9.1
asttokens==3.0.0
beautifulsoup4==4.14.3
cachetools==6.2.4
certifi==2025.11.12
charset-normalizer==3.4.4
colorama==0.4.6
decorator==5.2.1
distro==1.9.0
dj-rest-auth==7.0.1
Django==5.2.6
django-allauth==64.0.0
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
et_xmlfile==2.0.0
executing==2.2.1
finance-datareader==0.9.96
google-auth==2.45.0
google-genai==1.56.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
ipython==9.5.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
lxml==6.0.2
matplotlib-inline==0.1.7
narwhals==2.13.0
numpy==2.3.5
openpyxl==3.1.5
packaging==25.0
pandas==2.3.3
parso==0.8.5
pillow==12.0.0
plotly==6.5.0
prompt_toolkit==3.0.52
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
requests==2.32.5
requests-file==3.0.1
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
soupsieve==2.8
sqlparse==0.5.3
stack-data==0.6.3
tenacity==9.1.2
tqdm==4.67.1
traitlets==5.14.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.6.0
wcwidth==0.2.14
websockets==15.0.1
